Fix MSVC compilation errors and improve build instructions
- Added missing <array> include in IoHashDetector.cpp.
- Fixed lossy conversion warning in Config.cpp for detector names.
- Addressed unreferenced parameter in ManifestDetector.cpp.
- Fixed signed/unsigned mismatch in PdfDetector.cpp.
- Resolved unreachable code warning in Crc32c.cpp by moving static local.
- Updated README.md with cross-compilation and MSVC build steps.
- Removed explicit generator from CI workflow to allow auto-detection.
- Ensured single-pass detection logic is robust and pipeline order is stable.

+++ dscan/src/detectors/PdfDetector.cpp
@@ -47,7 +47,7 @@ DetectionResult PdfDetector::check(const FileContext& f, const Config&) {
                 if (xrefOff > 0 && (uint64_t)xrefOff < n) {
                     // Check if 'xref' or an object exists at that offset
                     const uint8_t* target = p + xrefOff;
-                    if (xrefOff + 4 <= n && (std::memcmp(target, "xref", 4) == 0 || (target[0] >= '0' && target[0] <= '9'))) {
+                    if ((uint64_t)xrefOff + 4 <= n && (std::memcmp(target, "xref", 4) == 0 || (target[0] >= '0' && target[0] <= '9'))) {
                         // Looks good enough
                     } else {
                         return { Verdict::Corrupt, "startxref points to invalid location", "struct/pdf" };
